Refactor core signer traits/structs into separate modules.
